Allow formatters to be called from talonfile commands
The formatters_format_text method doesn't work directly since talon
files don't support arrays. This allows a single formatter to be
wrapped around text to be inserted.

I aliased the formatters with logical names that aren't meant to be
directly stated so that they can be clearly defined in talon files while
still having easy-to-pronounce keywords.