Modified create spoken form to fix todos and add some other features (#1159)
This is a modified version of create_spoken_form.py that I worked on a
while ago, but am finally finding time to post it now. I've been using
it for quite a while and it seems to be fine, but it would be good for
other people to test. I left a couple commented out code in just to
explain why I changed it, but we could remove it from the final version
if folks are ok with the changes otherwise.

The main differences are: 
- When abbreviated versions are expanded, the original version is also
kept.
- When uppercase words are exploded, the original upper case and a new
lower case version are kept (ex: `LICENSE` becomes speakable as
`license` instead of only becoming `L I C E N S E`).
- Small lower case words (2-3 characters) are exploded into upper case
versions, under the assumption that they might be dictated by character
(ex: `usb` becomes `U S B`).
- Some words that are transformed are re-passed through other
transformation passes, this way if we do something like have an upper
case version of a word that's exploded, we may end up also passing the
lower case version through an reverse abbreviation pass.

The main thing I use this for is filesystem stuff in my shell, because I
have a zsh script that auto updates talon with files and folders in the
current directory. So I wanted ways of quickly being able to dictate
`vm`, `usb`, `README.md`, etc and it wasn't working well with the
existing version.

I didn't run the existing create_spoken_form tests that exist yet, and I
didn't add new tests. If someone wouldn't mind to quickly tell me what
the best way to run the tests are, I'm happy to do it and I can add new
ones for the stuff I added.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>