lang: Add tag for brace-delimited code blocks (#1332)
## Problem

Several languages had the same, or similar, copy-pasted implementations
of `code_block` to insert braces and leave the cursor on a blank line
between them. These all assume the editor will appropriately expand
<kbd>enter</kbd> between brackets, which isn't true for all editors.

The C# implementation implemented the same idea, but appears to assume
<kbd>enter</kbd> _isn't_ expanded.

Furthermore, some C-like languages were missing a `code_block` impl
entirely.

## Solution

Similar to #720, this adds a common implementation for braced code
blocks, which can be overridden on a per-editor basis, and adds the tag
to the appropriate languages.