Catch redundant sleep/wake commands, and notify the user that Talon is already in the requested mode (#1327)
This commit adds some commands to catch redundant sleep/wake commands,
and notifies the user.
The main example is, if Talon is already awake, the "talon wake" and
"wake up" commands aren't active. So if the user is unsure of their
current mode and says "wake up", Talon interprets the command as
something else ("page up", for example).
These new commands are active in the mode the user is trying to
activate. They catch the command, and trigger an `app.notify` to let the
user know they're already in that mode.

Before https://github.com/talonhub/community/pull/1215, these were bound
with `mode: all`, so using them in the wrong mode would still be
recognized, but just no-op.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Phil Cohen <phillip@phillip.io>
Co-authored-by: Jeff Knaus <knaus.jeff@gmail.com>